requirements.txt
discord.py
python-dotenv
aiohttp
pymongo
dnspython
replit